% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound-infinite-edges.R
\name{bound_infinite_edges}
\alias{bound_infinite_edges}
\title{Find boundary intersectinons of infinite edges}
\usage{
bound_infinite_edges(xmin, ymin, xmax, ymax, x, y, a, b, c, line, v1, v2)
}
\arguments{
\item{xmin, ymin, xmax, ymax}{bounding box to trim to.  Must be larger than 
bounding box of (x, y) points}

\item{x, y}{voronoi vertices}

\item{a, b, c}{line defintion ax + by = c}

\item{line, v1, v2}{voronoi segment definition: line, vertex1, vertex2}
}
\value{
list of new vertices and new edges
}
\description{
Find boundary intersectinons of infinite edges
}
