% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunes.R
\name{delaunay}
\alias{delaunay}
\title{Delaunay Triangulation}
\usage{
delaunay(x, y, calc_polygons = TRUE, calc_areas = FALSE, calc_segments = FALSE)
}
\arguments{
\item{x, y}{coordinates of seed sites. Duplicate points are not allowed.}

\item{calc_polygons}{calculate polygon coordinates default: TRUE}

\item{calc_areas}{calculate the areas of each triangle. Default: FALSE}

\item{calc_segments}{calculate segments. Default: FALSE}
}
\value{
named list of data
\describe{
 \item{sites}{coordinates of sites}
 \item{ntris}{Number of triangles in this delaunay triangulation} 
 \item{tris}{data.frame. Each row specifies the indices of the three sites
        which define a single delaunay triangle. If \code{calc_areas = TRUE} then
        this data.frame also includes an \code{area} column}
 \item{polygons}{data.frame of polygon coordinates with x,y and 
       polygon index. Only calculated when \code{calc_polygons = TRUE}}
 \item{segments}{The end-points and length of each unique undirected 
       edge in the triangulation. Only calculated when \code{calc_segments = TRUE}
    \describe{
       \item{v1,v2}{vertex indices referencing the original \code{sites}}
       \item{x1,y1,x2,y2}{coordinates of ends of segment}
       \item{len}{length of this segment}
    }      
 } 
}
}
\description{
Delaunay Triangulation
}
\examples{
set.seed(1)
x <- runif(10)
y <- runif(10)
del <- delaunay(x, y)
plot(del)
del
}
