% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_bounded_polygons}
\alias{draw_bounded_polygons}
\title{Add voronoi bounded polygons to an existing plot}
\usage{
draw_bounded_polygons(vor, ...)
}
\arguments{
\item{vor}{object returned by \code{voronoi()}}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None.
}
\description{
Add voronoi bounded polygons to an existing plot
}
