% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_finite_segments}
\alias{draw_finite_segments}
\title{Add voronoi segments to an existing plot}
\usage{
draw_finite_segments(vor, ...)
}
\arguments{
\item{vor}{object returned by \code{voronoi()}}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None.
}
\description{
Add voronoi segments to an existing plot
}
