% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-polygons.R
\name{draw_inf_lines}
\alias{draw_inf_lines}
\title{Add voronoi unbounded lines to an existing plot}
\usage{
draw_inf_lines(vor, lty = 1, col = "grey90", ...)
}
\arguments{
\item{vor}{object returned by \code{voronoi()}}

\item{lty}{linetype. default: 1}

\item{col}{colour. default: 'grey90'}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None.
}
\description{
Add voronoi unbounded lines to an existing plot
}
