% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi-plot.R
\name{draw_inf_segments}
\alias{draw_inf_segments}
\title{Add voronoi unbounded segments to an existing plot}
\usage{
draw_inf_segments(vor, col = "grey90", ...)
}
\arguments{
\item{vor}{object returned by \code{voronoi()}}

\item{col}{segment colour. Default: grey90}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None.
}
\description{
Add voronoi unbounded segments to an existing plot
}
