% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_polygons}
\alias{draw_polygons}
\title{Add voronoi bounded polygons to an existing plot}
\usage{
draw_polygons(vor, cols = rainbow(length(vor$polygons)), ...)
}
\arguments{
\item{vor}{object returned by \code{voronoi()}}

\item{cols}{vector of colours}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None.
}
\description{
Add voronoi bounded polygons to an existing plot
}
