% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_semi_infinite_segments}
\alias{draw_semi_infinite_segments}
\title{Add semi-infinite voronoi segments to an existing plot}
\usage{
draw_semi_infinite_segments(vor, col = "grey90", ...)
}
\arguments{
\item{vor}{object returned by \code{voronoi()}}

\item{col}{segment colour. Default: grey90}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None.
}
\description{
Add semi-infinite voronoi segments to an existing plot
}
