% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_sites}
\alias{draw_sites}
\title{Draw vertex nodes}
\usage{
draw_sites(vor, pch = "+", cex = 0.5, text = FALSE, ...)
}
\arguments{
\item{vor}{vor}

\item{pch}{plot character}

\item{cex}{point size scale}

\item{text}{show text?}

\item{...}{other arguments passed through}
}
\description{
Draw vertex nodes
}
