% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_vertices}
\alias{draw_vertices}
\title{Draw vertex nodes}
\usage{
draw_vertices(vor, nudge_x = 0.1, nudge_y = nudge_x, ...)
}
\arguments{
\item{vor}{vor}

\item{nudge_x, nudge_y}{text adjustment}

\item{...}{other arguments passed through}
}
\description{
Draw vertex nodes
}
