% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-polygons.R
\name{extract_polygons_r}
\alias{extract_polygons_r}
\title{Extract all bounded polygons from a voronoi tessellation (R version)}
\usage{
extract_polygons_r(vor, verbosity = 0)
}
\arguments{
\item{vor}{result of \code{voronoi()}}

\item{verbosity}{verbosity level. default: 0}
}
\value{
list of polygons
}
\description{
Extract all bounded polygons from a voronoi tessellation (R version)
}
