% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-vertices.R
\name{merge_vertices}
\alias{merge_vertices}
\title{Merge vertices}
\usage{
merge_vertices(x, y, line, v1, v2, tol = 1e-05, verbosity = 0L)
}
\arguments{
\item{x, y}{vertex coordinates}

\item{line}{edges defined by line two indices into the x, y coordinates}

\item{v1, v2}{edges defined by line two indices into the x, y coordinates}

\item{tol}{tolerance for merging close vertices.}

\item{verbosity}{verbosity level. Default: 0}
}
\value{
\code{list(line, v1, v2)}
}
\description{
Merge vertices
}
