% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-delaunay.R
\name{plot.del}
\alias{plot.del}
\title{Simple plot of Delaunay triangulation}
\usage{
\method{plot}{del}(
  x,
  sites = TRUE,
  tris = TRUE,
  segments = FALSE,
  site_pch = 19,
  site_col = "black",
  segment_col = "black",
  tri_col = rainbow(x$ntris),
  ...
)
}
\arguments{
\item{x}{output from \code{delaunay()}}

\item{sites, tris, segments}{logical. draw geometric feature}

\item{site_pch, site_col}{graphics parameters for each input site}

\item{segment_col}{colour for segments (length = 1)}

\item{tri_col}{colour for polygons (length = 1 or N)}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None
}
\description{
Simple plot of Delaunay triangulation
}
