% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vor}
\alias{plot.vor}
\title{Plot a Voronoi tessellation}
\usage{
\method{plot}{vor}(
  x,
  pch_site = ".",
  pch_vertex = 19,
  col_polygons = rainbow(length(vor$polygons)),
  buffer = 0,
  ...
)
}
\arguments{
\item{x}{object returned by \code{voronoi()}}

\item{pch_site}{plot character for sites}

\item{pch_vertex}{plot character for Voronoi vertices}

\item{col_polygons}{vector of colours for polygons}

\item{buffer}{buffer around extents. defualt: 0}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None.
}
\description{
Barebones plotting of Voronoi
}
