% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vor}
\alias{plot_vor}
\title{Plot the vertices of the voronoi tessellation}
\usage{
plot_vor(vor, buffer = 0, asp = 1, ann = FALSE, axes = FALSE, pch = ".", ...)
}
\arguments{
\item{vor}{object returned by \code{voronoi()}}

\item{buffer}{buffer around extents. defualt: 0}

\item{asp}{aspect ration. Default 1}

\item{ann}{annotation? Default: FALSE}

\item{axes}{draw axes? Default: FALSE}

\item{pch}{plot character}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
None.
}
\description{
Plot the vertices of the voronoi tessellation
}
