% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-matching.R
\name{point_in_polygon}
\alias{point_in_polygon}
\title{Point in convex polygon?}
\usage{
point_in_polygon(x, y, xp, yp)
}
\arguments{
\item{x, y}{point}

\item{xp, yp}{polygon vertices}
}
\value{
logical
}
\description{
Point in convex polygon?
}
