% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-grid.R
\name{voronoiGrob}
\alias{voronoiGrob}
\title{Plot voronoi}
\usage{
voronoiGrob(vor, fill = rainbow(length(vor$polygons)))
}
\arguments{
\item{vor}{voronoi object as returned by \code{voronoi()}}

\item{fill}{fill colours for polygons}
}
\value{
None
}
\description{
Plot voronoi
}
