% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunes.R
\name{delaunay}
\alias{delaunay}
\title{Calculate Voronoi}
\usage{
delaunay(x, y, calc_polygons = TRUE)
}
\arguments{
\item{x, y}{coordinates of seed points}

\item{calc_polygons}{calculate and return polygon coordinates? default: TRUE}
}
\value{
data.frame. Each row specifies the indices of the three seed points
        which define a single delaunay triangle
}
\description{
Calculate Voronoi
}
\examples{
set.seed(1)
x <- runif(10)
y <- runif(10)
delaunay(x, y)
}
