% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunes.R
\name{delaunay}
\alias{delaunay}
\title{Delaunay Triangulation}
\usage{
delaunay(x, y, calc_polygons = TRUE)
}
\arguments{
\item{x, y}{coordinates of seed sites. Duplicate points are not allowed.}

\item{calc_polygons}{calculate and return polygon coordinates? default: TRUE}
}
\value{
named list of data
\describe{
 \item{segments}{data.frame. Each row specifies the indices of the three seed points
        which define a single delaunay triangle}
 \item{polygons}{data.frame of polygon coordinates}
}
}
\description{
Delaunay Triangulation
}
\examples{
set.seed(1)
x <- runif(10)
y <- runif(10)
delaunay(x, y)
}
