% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-polygons.R
\name{extract_polygons}
\alias{extract_polygons}
\title{Extract polygons}
\usage{
extract_polygons(x, y, v1, v2, verbosity = 0L)
}
\arguments{
\item{x, y}{vertex coordinates}

\item{v1, v2}{edges defined by two indices into the x, y coordinates}

\item{verbosity}{verbosity level. Default: 0}
}
\value{
\code{list(list(x = ..., y = ...), list(x = ..., y = ...), ...)}
}
\description{
Extract polygons
}
