% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunes.R
\name{voronoi}
\alias{voronoi}
\title{Voronoi tessellation on an infinite plane}
\usage{
voronoi(x, y, calc_polygons = TRUE, match_sites = TRUE)
}
\arguments{
\item{x, y}{coordinates of seed points}

\item{calc_polygons}{calculate polygons from the voronoi vertices and segments
Default: TRUE}

\item{match_sites}{logical. SHould the polygons be re-ordered to match 
the seed points? Default: TRUE}
}
\value{
named list of data.frames.
\describe{
  \item{vertex}{data.frame of vertex coordinates in the tessellation}
  \item{line}{data.frame of line equations in the tessellation of the form 'ax + by = c'}
  \item{segment}{data.frame of segments defined by 'line', 'v1' and 'v2'.  'line'
        is the row index into the 'line' data.frame.  'v1' and 'v2' are 
        the indices into 'vertex' which define the endpoints along the 
        specified 'line'.  If 'v1' or 'v2' are NA this indicates that the 
        segment continues to infinity}
}
}
\description{
Voronoi tessellation on an infinite plane
}
\examples{
set.seed(1)
x <- runif(10)
y <- runif(10)
voronoi(x, y)
}
