% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunes.R
\name{voronoi}
\alias{voronoi}
\title{Voronoi Tessellation}
\usage{
voronoi(
  x,
  y,
  calc_polygons = TRUE,
  match_sites = TRUE,
  bound_segments = TRUE,
  merge_tolerance = 1e-10
)
}
\arguments{
\item{x, y}{coordinates of seed sites. Duplicate points are not allowed.}

\item{calc_polygons}{Logical. Should voronoi polygons be calculated?
Default: TRUE}

\item{match_sites}{Logical. Should the polygons be re-ordered to match 
the seed points? Default: TRUE. This option only makes sense
when \code{calc_polygons = TRUE}.}

\item{bound_segments}{logical. Default: TRUE.  If \code{calc_polygons = TRUE}
then this option is always TRUE.}

\item{merge_tolerance}{Limit of how close the ends of a segment must be
before the segment is collapsed to a single vertex Default: 1e-10}
}
\value{
An object of class "vor" which is a named list of data.frames.
\describe{
  \item{sites}{data.frame of original sites}
  \item{vertices}{data.frame of voronoi vertices. This is the raw output from Fortune's algorithm}
  \item{segments}{data.frame of segments defined by 'line', 'v1' and 'v2'.  'line'
        is the row index into the 'lines' data.frame.  'v1' and 'v2' are 
        the indices into 'vertices' which define the endpoints along the 
        specified 'line'.  If 'v1' or 'v2' are NA this indicates that the 
        segment continues to infinity}
  \item{polygons}{
     polygon information for each voronoi cell. Only calculated when
     \code{calc_polygons = TRUE}
     \describe{
       \item{npolygons}{Number of polygons}
       \item{coords}{data.frame of x,y coordiantes and polygon index for all polygons}
       \item{bbox}{data.frame of polygon bounding box information. Each row represents a polygon}
       \item{centrois}{data.frame of polygon centroids. Each row represents a polygon}
     }
  }
  \item{lines}{data.frame of line equations in the voronoi diagram of the form 'ax + by = c'}
  \item{extents}{list of (xmin, ymin), (xmax, ymax) bounding box which encompasses
        all input sites and voronoi vertices}
  \item{mvertices}{Merged and bounded vertices used to calculate polygons} 
  \item{msegments}{Merged and bounded segments used to calculate polygons} 
}
}
\description{
Voronoi Tessellation
}
\examples{
set.seed(1)
x <- runif(10)
y <- runif(10)
vor <- voronoi(x, y)
plot(vor)
vor
}
